renamed post folder to  projects
